pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "citynexus"
version = "0.1.0"
description = "Self-evolving multi-agent urban simulation environment"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "openenv-core>=0.2.0",
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0",
]

[tool.setuptools.packages.find]
where = ["src"]
